objectified fitness function. added slim fitness for EHC updates. Worked on memory management with custom swap functions for struct ind and optimizing sort functions.

+++ DevelEP/rnd.h
@@ -39,11 +39,11 @@ public:
 		float fac,rsq,v1,v2;
 		if (iset == 0) {// We don't have an extra deviate handy, so 
 			do{
-				v1=2.0*rnd_flt(-1,1)-1.0; //pick two uniform numbers in the square ex
-				v2=2.0*rnd_flt(-1,1)-1.0; //tending from -1 to +1 in each direction,
+				v1=float(2.0*rnd_flt(-1,1)-1.0); //pick two uniform numbers in the square ex
+				v2=float(2.0*rnd_flt(-1,1)-1.0); //tending from -1 to +1 in each direction,
 				rsq=v1*v1+v2*v2;	   //see if they are in the unit circle,
 			} while (rsq >= 1.0 || rsq == 0.0); //and if they are not, try again.
-			fac=sqrt(-2.0*log(rsq)/rsq);
+			fac=float(sqrt(-2.0*log(rsq)/rsq));
 		//Now make the Box-Muller transformation to get two normal deviates. Return one and
 		//save the other for next time.
 		gset=v1*fac;
